add library flow, organize login files
