+++ maestro/tests/library.yaml
@@ -1,0 +1,17 @@
+appId: com.jellify
+---
+# Wait for app to be ready, then navigate to Settings tab
+- assertVisible: 
+    id: "library-tab-button"
+
+# Navigate to Library tab using text
+- tapOn:
+    id: "library-tab-button"
+
+# Verify we're on the library page
+- assertVisible:
+    id: "library-artists-tab-button"
+
+# Navigate to Artists tab
+- tapOn:
+    id: "library-artists-tab-button"
