+++ maestro/tests/settings.yaml
@@ -3,7 +3,6 @@ appId: com.jellify
 # Wait for app to be ready, then navigate to Settings tab
 - assertVisible: 
     id: "settings-tab-button"
-#    timeout: 5000
 
 # Navigate to Settings tab using text
 - tapOn:
@@ -12,7 +11,6 @@ appId: com.jellify
 # Verify we're on the settings page
 - assertVisible:
     text: "App"
-#    timeout: 3000
 
 # Test App (Preferences) Tab - should already be selected
 - assertVisible:
