+++ src/screens/Login/server-address.tsx
@@ -1,23 +1,23 @@
 import React, { useEffect, useState } from 'react'
 import _, { isUndefined } from 'lodash'
 import { useMutation } from '@tanstack/react-query'
-import { JellifyServer } from '../../../types/JellifyServer'
+import { JellifyServer } from '../../types/JellifyServer'
 import { Input, ListItem, Separator, Spinner, XStack, YGroup, YStack } from 'tamagui'
-import { SwitchWithLabel } from '../../Global/helpers/switch-with-label'
-import { H2, Text } from '../../Global/helpers/text'
-import Button from '../../Global/helpers/button'
-import { http, https } from '../utils/constants'
+import { SwitchWithLabel } from '../../components/Global/helpers/switch-with-label'
+import { H2, Text } from '../../components/Global/helpers/text'
+import Button from '../../components/Global/helpers/button'
+import { http, https } from '../../components/Login/utils/constants'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
-import { StackParamList } from '../../types'
+import { StackParamList } from '../../components/types'
 import Toast from 'react-native-toast-message'
-import { useJellifyContext } from '../../../providers'
-import { useSettingsContext } from '../../../providers/Settings'
-import Icon from '../../Global/components/icon'
+import { useJellifyContext } from '../../providers'
+import { useSettingsContext } from '../../providers/Settings'
+import Icon from '../../components/Global/components/icon'
 import { PublicSystemInfo } from '@jellyfin/sdk/lib/generated-client/models'
-import { connectToServer } from '../../../api/mutations/login'
-import { IS_MAESTRO_BUILD } from '../../../configs/config'
-import { sleepify } from '../../../utils/sleep'
+import { connectToServer } from '../../api/mutations/login'
+import { IS_MAESTRO_BUILD } from '../../configs/config'
+import { sleepify } from '../../utils/sleep'
 
 export default function ServerAddress({
 	navigation,
