+++ src/screens/Login/server-authentication.tsx
@@ -1,19 +1,19 @@
 import React, { useState } from 'react'
 import { useMutation } from '@tanstack/react-query'
 import _ from 'lodash'
-import { JellyfinCredentials } from '../../../api/types/jellyfin-credentials'
+import { JellyfinCredentials } from '../../api/types/jellyfin-credentials'
 import { getToken, H6, Spacer, Spinner, XStack, YStack } from 'tamagui'
-import { H2, H5, Text } from '../../Global/helpers/text'
-import Button from '../../Global/helpers/button'
+import { H2, H5, Text } from '../../components/Global/helpers/text'
+import Button from '../../components/Global/helpers/button'
 import { SafeAreaView } from 'react-native-safe-area-context'
-import { JellifyUser } from '../../../types/JellifyUser'
-import { StackParamList } from '../../types'
-import Input from '../../../components/Global/helpers/input'
-import Icon from '../../Global/components/icon'
-import { useJellifyContext } from '../../../providers'
+import { JellifyUser } from '../../types/JellifyUser'
+import { StackParamList } from '../../components/types'
+import Input from '../../components/Global/helpers/input'
+import Icon from '../../components/Global/components/icon'
+import { useJellifyContext } from '../../providers'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
 import Toast from 'react-native-toast-message'
-import { IS_MAESTRO_BUILD } from '../../../configs/config'
+import { IS_MAESTRO_BUILD } from '../../configs/config'
 
 export default function ServerAuthentication({
 	navigation,
