+++ src/screens/Login/server-library.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { NativeStackNavigationProp } from '@react-navigation/native-stack'
-import { StackParamList } from '../../../components/types'
-import { useJellifyContext } from '../../../providers'
+import { StackParamList } from '../../components/types'
+import { useJellifyContext } from '../../providers'
 import { BaseItemDto } from '@jellyfin/sdk/lib/generated-client/models'
-import LibrarySelector from '../../Global/components/library-selector'
+import LibrarySelector from '../../components/Global/components/library-selector'
 
 export default function ServerLibrary({
 	navigation,
